backend/requirements.txt
# Python dependencies - to be filled from requirements_txt.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Google Gemini API
google-generativeai==0.3.0

# GitHub API
PyGithub==2.1.1
httpx==0.25.2

# Caching and Data
redis==5.0.0
python-dotenv==1.0.0

# Async Support
asyncio==3.4.3

# Logging and Utilities
rich==13.7.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Type Checking
mypy==1.7.0

# Code Quality
black==23.12.0
ruff==0.1.8